nixpkgs-index/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nixpkgs-index"
version = "0.1.0"
description = "Index package versions across nixpkgs commits"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "colorlog>=6.7.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
nixpkgs-index = "nixpkgs_index.cli:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
